removed images loading and addedd loading component. replaced button with search bar needs search button logic implemented needs search bar logic implemented against properties db
