--- lendlogic-app/app/components/NewBuyerQuestions.jsx
@@ -1,9 +1,0 @@
-
-
-export default function NewBuyerQuestions() {
-  return (
-    <div>
-      
-    </div>
-  )
-}
