--- lendlogic-app/app/components/RemortgageLearningTool.jsx
@@ -1,9 +1,0 @@
-
-
-export default function RemortgageLearningTool() {
-  return (
-    <div>
-      
-    </div>
-  )
-}
