+++ lendlogic-app/app/dashboard/newbuyer_components/ExpandChecklist.jsx
