+++ lendlogic-app/app/dashboard/newbuyer_components/NewBuyerTimeline.jsx
@@ -1,0 +1,12 @@
+import Step from "../Step";
+
+export default async function NewBuyerTimeline({ steps }) {
+  return (
+    <div>
+      <h2>Gamefied Buyer Timeline</h2>
+      {/* {steps.map((step) => (
+        <Step key={step.id} title={step.title} tasks={step.tasks} />
+      ))} */}
+    </div>
+  );
+}
