+++ lendlogic-app/app/dashboard/page.jsx
@@ -1,15 +1,12 @@
 import LearningSection from "./LearningSection";
 import Newsletter from "./Newsletter";
-import RemortgageReport from "./RemortgageReport";
+import RemortgageReport from "./remortgage_components/RemortgageReport";
 import DisplayUser from "../customHooks/DisplayUser";
 import DisplayProperties from "../customHooks/DisplayProperties";
-import NewBuyerReport from "./NewBuyerReport";
-import { Suspense } from "react";
-
-import Image from "next/image";
+import NewBuyerReport from "./newbuyer_components/NewBuyerReport";
 import { cookies } from "next/headers";
-import Checklist from "./Checklist";
-import BuyingHomeTimeline from "./BuyingHomeTimeline";
+import Checklist from "./newbuyer_components/Checklist";
+import BuyingHomeTimeline from "./newbuyer_components/BuyingHomeTimeline";
 import GetSteps from "../customHooks/DisplaySteps";
 
 export default async function Dashboard() {
@@ -43,7 +40,6 @@ export default async function Dashboard() {
         <p className="px-2 py-4 text-lg">
           Welcome {currentUser.userName}, here is everything you need to know
         </p>
-
         {q1 === "a2" ? <RemortgageReport value={currentUser} q2={q2} /> : null}
         {q1 === "a1" ? (
           <NewBuyerReport value={currentUser} properties={properties} />
